NewsNoticeEditor 분리
